+++ src/governance-compliance/service-registration.js
@@ -1,6 +1,7 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { GovernanceOperatingModelDescriptor } from './operating-model/governance-operating-model-descriptor.js';
 import { GovernanceComplianceOverviewDescriptor } from './overview/governance-compliance-overview-descriptor.js';
+import { PolicyArchitectureDescriptor } from './policy-architecture/policy-architecture-descriptor.js';
 
 export function addGovernanceCompliance(services) {
   if (!(services instanceof ServiceCollection)) {
@@ -9,6 +10,7 @@ export function addGovernanceCompliance(services) {
 
   services.registerSingleton('GovernanceComplianceOverviewDescriptor', () => new GovernanceComplianceOverviewDescriptor());
   services.registerSingleton('GovernanceOperatingModelDescriptor', () => new GovernanceOperatingModelDescriptor());
+  services.registerSingleton('PolicyArchitectureDescriptor', () => new PolicyArchitectureDescriptor());
 
   return services;
 }